requirements.txt
azure-mgmt-resource>=21.1.0
azure-identity>=1.12.0
azure-ai-ml>=1.9.0
azure-mgmt-ml>=1.0.0
openai>=1.3.0
python-dotenv>=1.0.0